Superior String Fish
This is a highly successful version of the StringFish. Ironically, despite being created with the idea of using a String to minimize calls on info(), it became easier to think about decisions in terms of constant calls. There's not a penalty for creating so many calls, so I have not used it. However, the structure to support the string building helped develop the friendCount and enemyCount, two variables fundamental to the decisional tree of this critter. 
It cannot yet get bored of its groups and seek new ones. This will be added. I also hope to add scholastic movement, if I discover a way to move large groups all at once.